eglGetFrameTimestamps: Don't require reads done to be monotonic.
Reads done is usually monotonic, but only under the assumption
that every buffer of the surface is being consumed by the same
unit. It is possible for a surface to switch between being an
overlay and being composited by the GPU. It's also possible a
driver will maintain a shadow copy of a frame that will further
affect when reads from a buffer might finish.

Remove the overly strict check.

Bug: 70334851, 74184639
Test: --deqp-case=dEQP-EGL*get_frame_timestamps*
Change-Id: I8e242a9fd649720b4d8059797c71163329b7b92c