Enable shader cache by default
This change flips the command line option for the shader cache to be
enabled by default.

The change also incorporates shader cache truncation, which is enabled
by default. This resets the shader cache on the start of a CTS run. For
more performance the option can be disabled so the shader cache is used
across CTS runs, but this may lead to very large shader cache files over
time.

Affects:

dEQP-VK.*

Components: Framework, Vulkan

VK-GL-CTS issue: 899

Change-Id: Ia1bc8c932b6b280e2815f7de7f6e715b0e4edb27