Integrated SPIRV-Tools-opt to conformance test
Added optional shader optimization pass to all shaders in the
conformance test. The optimizer is disabled by default, but can
be enabled through a command-line parameter.

Due to the nature of optimizations and certain hand-crafted
SPIR-V tests, the optimizer MUST NOT be enabled by default, because
it will alter the tests and possibly eliminate the things that
are being tested.

Running with the optimizer for testing is, however, recommended
since it can expose bugs in drivers. Only some optimization recipes
are currently supported, but adding new ones is relatively simple.

Affects:

None by default, most tests if enabled.

Components: Framework, Vulkan

VK-GL-CTS issue: 92

Change-Id: I74996448fda1f4e11a9b16bba8c44561723a1e4e