Add test for VK_EXT_shader_stencil_export
The extension allows fragment shader to output a reference value for
the pixel. The test uses a first subpass to set reference values based
on position and VK_STENCIL_OP_REPLACE to write those to the stencil
buffer. Then a second subpass will draw to a color buffer using the
generated stencil.

A new overload for addFunctionCaseWithPrograms() was added to the
framework.

New tests:
    dEQP-VK.pipeline.shader_stencil_export.op_replace

Components: Framework, Vulkan

VK-GL-CTS issue: 1029

Change-Id: Ic2c2f9ebe33c8575e1d6ccb20b2435aadc999b84