Added test to check memory requirement invariance
Added a test that allocates random buffers and images, first in order
one by one, and then shuffled, verifying that the required memory
sizes match between these runs.

Image sizes, formats, allocation types and heap type requirements are
randomized.

The test depends on the supported image formats and available heap
types, so it is subtly different depending on the features the
driver supports. The goal for this test is to check for memory
requirement invariance, and thus it differs with the exact features
that are supported. The used features are printed in the log
for reference.

New tests:

dEQP-VK.api.invariance.random

Components: Vulkan

VK-GL-CTS issue: 1171

Change-Id: If869a72f6552174171d1f9bcce51e424ad8e8572