Search a longer list of random triangles in EGL render tests
This set of tests contain a loop that searches a random set of
coordinates for triangles that are large enough.

The random number seed used is the test name, and it happens that
this test's seed does not generate enough large triangles in under
40 iterations (the max before abort)

This change increases the maximum number of iterations to 100.
This will have no impact on runtime/output on any tests that
already successfully run.

Affects:
dEQP-EGL.functional.render.*

Components:
dEQP

VK-GL_CTS issue:
1213

Change-Id: Ibbd8316940f7d43d343082d21ae7aa7d17d865b4