Separate out debug information in Android build
We don't want to place the debug information in the APK, but we should
keep it around on the host so that the debugger can work.

In a "normal" Android app packaging process this is done for us by
gradle, but we roll our own. Do it as a CMake post-build step instead.

Change-Id: I59b13659d66333a06ce60b6057db19b39c067f53
Components: Framework
VK-GL-CTS: 1194
VK-GL-CTS: 1533
(cherry picked from commit 651bd0d1bcfc3763521f16bfd3b9a40396998e8c)