Fix validation errors in draw tests
Many of the draw tests were causes validation errors
because the VkMappedMemoryRange.size parameter was not
a multiple of VkPhysicalDeviceLimits::nonCoherentAtomSize.

In all cases for these tests the size can be set to
VK_WHOLE_SIZE to avoid the validation error. VK_WHOLE_SIZE
can be used since in all cases the test flush the whole
mapped range.

Affects: dEQP-VK.draw.*

Change-Id: Idfa664ed03a1b0f55b356a443a7bc309ca7fae20
Components: Vulkan