Robust_buffer_access_behavior.texel_fetch relies on undefined behavior
The failed test behavior is :
Create a mipmap texture as a dst texture and attach its level 1 texture
to the framebuffer colorattachment0 as render target. Create a mipmap
texture as a src texture and fill some data to its level 1 texture.
Use "texelFetch" in the fragment shader to sample level 1 source
texture(mipmap) and draw it to the dst texture.

According to the GLES3.2 spec:
The results of texelFetch builtins are undefined if the computed level
of detail is not the texture's base level and the texture's minification
filter is NEAREST or LINEAR.

Modify the test for the mipmap case.

Components: OpenGL

Affects:
KHR-NoContext.es32.robust_buffer_access_behavior.texel_fetch

VK-GL-CTS issue: 1236

Change-Id: Id67ea894a3dd423bba611ccc11a131b80d38fc64