Add tests using VkRunner, precompiling the scripts
This is an alternative version of change 2660 that precompiles the
GLSL or SPIR-V assembly in the VkRunner test scripts to SPIR-V binary
during the build phase so that they can be executed faster and without
any external glslang or spirv-tools dependencies.

This adds VkRunner as an external dependency and then adds some
example test cases to use it. VkRunner is a tool designed to make it
easier to create shader tests for Vulkan by using a simple scripting
language which combines the shaders used for a pipeline along with a
description of the inputs and expected outputs.

VkRunner is fetched from an external repository so the fetch_sources
script needs to be run after merging the patch.

VkRunner tests can be added to any group, the tests need to be a child
class of the VkRunnerTestCase class.

Some VkRunner tests are provided as example, but they are excluded from
the mustpass:

   dEQP-VK.vkrunner-example.*

Shader test files are under external/vulkancts/data/vulkan/vkrunner/ path.
The example shader_test files are under example/ subfolder.

VK-GL-CTS: 1147
Change-Id: I1c5fc5f672a360800bfe10afbf96ca8fd07b38e6