Fix swap_buffers_with_damage.resize* hang on X11
In SwapBuffersWithDamageTest, the second time setSurfaceSize() was
called, because the size values were the same as the first call, the
test would hang with XNextEvent() in XlibWindow::setDimensions() waiting
for a ConfigureNotify event that Xserver would not deliver in
ConfigureWindow() if size was not changed.

Changing the if conditions causes size values to change for neighboring
setSurfaceSize calls and resolves the hang.

Affects:
dEQP-EGL.functional.swap_buffers_with_damage.resize*

Components: egl

VK-GL-CTS issue: 1238

Change-Id: I78177ffbef4f7f380d850ad43d83317f82401c7b