Max compiler threads only specifies a suggestion
In this test, it is expected that calling MaxShaderCompilerThreadsKHR()
with a count of 0 will compile non-parallel. However, the extension
purposefully says this is only a  "request" for non-parallel compiling
or linking.

This change will remove this specific test.

Affects:

parallel_shader_compile.compilation_completion_non_parallel

Components: OpenGL

VK-GL-CTS issue: 1226

Change-Id: I49214b3b32b4991ee7b96aad3853e24075c0f137