Fix dEQP-GLES3.functional.fbo.multiview.*
When the width of render target is an odd value, these tests failed.

The reason is these tests set default sample parameters and sample from
a 512x512 surface, where the left/right views have different colors.
When the width of render target is an odd value, the color of central
line is undefined.

As these tests' purpose is mainly want to verify whether the left/right
views have the different color, but not exactly care what the central
line is, a simple solution is just skip the check of central line.

Components: OpenGL

VK-GL-CTS issue: 1256

Affects: dEQP-GLES3.functional.fbo.multiview.*

Change-Id: I62c86dc7269864a3919b11a331178bfc54d65ab0