Test gl_FrontFacing with both GL_CCW and GL_CW.
glFrontFace(G_CW) should produce the inverse result of
glFrontFace(G_CCW):

OpenGL ES 3.0.5 spec:
"The built-in variable gl_FrontFacing is set to TRUE if the fragment is generated
from a front-facing primitive, and FALSE otherwise. For fragments generated
from triangle primitives, the determination is made by examining the sign of the
area computed by equation 3.6 of section 3.6.1 (including the possible reversal of
this sign controlled by FrontFace). If the sign is positive, fragments generated by
the primitive are front-facing; otherwise, they are back-facing."

Components: AOSP
Affects:
	dEQP-GLES2.functional.shaders.builtin_variable.frontfacing
	dEQP-GLES3.functional.shaders.builtin_variable.frontfacing
VK-GL-CTS Issue: 1334

Change-Id: I15d842055f789f8fa003e32eb70dbe3056c4b3f2