Add a flushAlloc helper to fix valid usage
Flushing/invalidating memory using the requested size (as the CTS code
does all over the place) is not safe. If the implementation's
MemoryRequirements make the allocation larger than the requested size
then the whole (padded) allocation will be mapped. This means that flush
and invalidate using a size which is not a multiple of the non-coherent
atom size are invalid. Replace these usages with helper functions to
flush/invalidate the whole block.

The current implementation uses WHOLE_SIZE to ensure that flush /
invalidate is valid, but sub-ranges could be implemented by storing how
much of any allocation is mapped. The structure of the current code
which maps memory at allocation time means that there cannot be any
other users being flushed out by the use of WHOLE_SIZE.

Includes a significant amount of work from @jbolz.

Components: Framework, Vulkan
Affects: dEQP-VK.*

Change-Id: I24ac185d7d041239430c53a6b61cede5828a7afd