Added script for vulkan.h.in generation
In order to avoid manual editing of vulkan.h.in, this script can be used
to perform all the operations that usually need manual editing, but
instead performs the operations from source XML files instead.

The script can read a base XML file, a superset XML file and the current
header file. These three files are analysed for the extensions they
contain, and the user can interactively select which extensions from the
superset to include in the generated header. (All extensions from the
base XML are included).

The script also automatically fetches the current vk.xml and header
genration scripts from github using wget.

This script is not automatically run by the build process, and requires
manual interaction to run. It depends on python3, since it uses the
header generation scripts that are part of the vk-docs project. Only the
generated header files should be committed to the git repository.

Affects:

None

Components: Framework

VK-GL-CTS issue: 1322

Change-Id: I26557d672483325924dc32febadd283669a4bc82