VK_EXT_memory_budget and VK_EXT_memory_priority
For VK_EXT_memory_budget, this is a simple query and sanity check
of the values in the new structure. There's not a whole lot about
this extension that can really be tested, since it is intentionally
very implementation-dependent. But this may be useful to implementers.

For VK_EXT_memory_priority, add variants of the memory.binding tests
that vary the priority. There's nothing really functional to test
here except that things still work when a priority is set.

Components: Vulkan

New Tests:
dEQP-VK.api.info.device.memory_budget
dEQP-VK.memory.binding.priority*

VK-GL-CTS issue: 1688
VK-GL-CTS issue: 1689

Change-Id: I376d2513f772051ae86eb7e9baa22eee68e3d39d