Fix SPIR-V validation in float_controls
FPRoundingMode decoration can be applied only to the Object
operand of an OpStore. This couldn't be done in vertex stage
without using write to SSBO which is not universally supported.
Because of that two vertex stage tests had to be removed.
FPRoundingMode is still tested in compute and fragment stage.

Components: Vulkan

VK-GL-CTS issue: 1430

Affects:
dEQP-VK.spirv_assembly.instruction.graphics.float_controls.*

Change-Id: I4f01149dae3ec9184b99302d51d681739ed15bf9