Added tests to intentionally mismatch opImageType
These tests intentionally mismatch the opImageType of the shader with
the source image types. The results of execution are ignored, but the
driver and/or graphics device is not expected to crash.

For the graphics tests, the source image format is changed, and
for the compute tests the shader is changed.

New tests:

dEQP-VK.spirv_assembly.instruction.graphics.image_sampler.*.optypeimage_mismatch.*
dEQP-VK.spirv_assembly.instruction.compute.image_sampler.*.optypeimage_mismatch_*

Components: Vulkan

VK-GL-CTS issue: 1420

Change-Id: I6f86986fda108b0dafc1a10d1790414d7bd5a2ba