Fix subgroup shuffle tests to do testing
Modify the subgroup shuffle tests to modify the indices so that they're
in range, rather than just rejecting all the values that were out of
range. In my testing of the old method, every test was skipped so the
tests always passed.

Components: Vulkan
VK-GL-CTS issue: 1490
Affects: dEQP-VK.subgroups.shuffle.*

Change-Id: I51d4511aef60c37f39c634be9dbeaff9aca4717e