Add subpass external dependency to pushDescriptor tests
Back to back render passes are used with no pipeline barrier
or external subpass dependency to synchronise the output of
the first to the input of the second.

This change adds an external subpass dependency to ensure that
the render passes are synchronised.

Component: Vulkan
Affects: dEQP-VK.pipeline.push_descriptor.*input_attachment

VK-GL-CTS issue: 1467

Change-Id: Ie42ec31ab393439e7a4c312cfe6dcc40fde82c25