Run validation before optimization.
The optimizer runs is own pass of the validator as a safety check.
However, that call to the validator does not get the any validatoin options passed in.
This causes problems with the memory layout tests.

The solution we will be to diable the call to the validator in the
optimizer.  Instead the will be an explict call to validate the binary
before optimization starts, reusing functions that already exist to do
that.

At the same time, the run of the validator that is done with
--deqp-validation=enable will be moved so that it runs after
optimization.  This way it will validate the output instead of the
input, and it will not be a redundant check.

VK-GL-CTS issue: 1503

Components: Vulkan, Framework

Affects: dEQP-VK.ssbo.layout.random.relaxed.*

Change-Id: Ib52fa97505855316739e07e472d23fcab4ac2beb