Truncate random coordinates in late fragment tests
The late-fragment tests generate random triangles.  Unfortunately,
the reference rasteriser does not respect the sub-pixel precision
of the physical device (it assumes 8 bits) which means that on
some platforms the reference rasterisation does not agree with
implementation.

This patch rounds the random coordinates to the reported
sub-pixel precision.  It was decided that in the short term this
was a less invasive change than modifying the reference rasteriser.

Affects: dEQP-VK.renderpass*.subpass_dependencies.late_fragment_tests.*

Components: Vulkan

VK-GL-CTS issue: 1549

Change-Id: I50f20284d9d2cb22e73564d6f71c0a3342547b68