Fix NaN handling in funord tests
Support of NaNs is optional according to the specification, but
these tests were assuming that NaNs are always supported.
Modify these tests to skip check if an operand is a NaN.

Add infrastructure for a new set of tests opfunord_nan*, which
use VK_KHR_shader_float_controls to check whether NaNs are supported.
But since this is a bugfix, do not actually add the new tests.

Affected tests:
* dEQP-VK.spirv_assembly.instruction.compute.opfunord*

Components: Vulkan

VK-GL-CTS issue: 1551
Change-Id: Ic60f13377ff9d2bdc2e10047aab8be0617904670