Add control for subpixel bits to reference rasterizer
The reference rasterizer assumed that subpixel bits was 8,
and this could lead to issues on some platforms.

This change makes it mandatory to specify the number of
subpixel bits required when using the reference rasterizer.

EGL, OpenGL and Vulkan tests that use the reference rasterizer
are changed to supply the platform-advertised subpixel bits.

Given the invasive nature of this change, the affects list
is extremely conservative.

Affects:
dEQP-VK*
dEQP-GL*
dEQP-EGL*

Components: Vulkan, OpenGL, EGL, Framework.

VK-GL-CTS issue: 1566

Change-Id: I6e89cc6e3b71b7d2d7fb02c4c21857ec4ad0146c