Add test for concurrent compute
Added test that runs identical payload on two queues with different
priority levels, and checking that the lower priority queue does not
finish first, adding a note to the log if the priorities seem to be
reversed. At the end it is also verified that both queues did the
work.

New test:

dEQP-VK.compute.basic.concurrent_compute

Components: Vulkan

VK-GL-CTS issue: 1509

Change-Id: Ib5d52dcbb3cee0e07c4d14bbb8d61d2f1ee2a8c5