Add alpha to coverage test with no color attachment
Added a test for alpha to coverage without a color
attachment. First the test draws a primitive with alpha 0
and a subpass with only a depth buffer. No writes to a depth
buffer are expected. Then a second draw with a color buffer
and the same depth buffer is done to verify the depth
buffer still has the original clear values.

New tests:

dEQP-VK.pipeline.multisample.alpha_to_coverage_no_color_attachment.*

Components: Vulkan

VK-GL-CTS issue: 1586
Change-Id: Ibfc9cbf0ade3280fe80f43296fad79b5c6f76ba1