Disable SPIR-V ids optimization
The default for the SPIR-V assembler is to compress IDs.
So even if you have %111111111 it will reassign that to the
next unused ID (a bump allocator). To disable this I need
to specify the equivalent of the --preserve-numeric-ids
option to the assembler. After discussion with Alex I specify
SPV_TEXT_TO_BINARY_OPTION_PRESERVE_NUMERIC_IDS for all tests.

Affected tests:
 * dEQP-VK.*

Components: vulkan

VK-GL-CTS issue: 1583

Change-Id: I941adf5d5cdb964eb4dff46afa75b01e089d3061