Fix validation errors in dEQP-VK.tessellation.*
In multiple tests memory allocations are invalidated or flushed using
a size that on some devices is not a multiple of nonCoherentAtomSize
device limit, causing validation errors. This fix uses helper functions
that invalidate or flush the whole allocation where the size of the
memory being invalidated/flushed equals the size of the allocation.

Affects:

dEQP-VK.tessellation.*

Components: Vulkan

VK-GL-CTS issue: 1620

Change-Id: I3c0aed7f749693808331a1b430ad2cb28b138699