Check for getPlatformDisplay support before using
I was under the impression that NativeDisplay::CAPABILITY_GET_DISPLAY_PLATFORM
meant that we know the capability exists. But that apparently isn't
true.
Instead treat that as a flag to enable checking for the capability.
Don't really see the point of the capability bit if I have to check
anyway.

Affects: Android
KHR-GLES32.info.*
Pretty much any test running on EGL 1.4 device.

Components: Framework, AOSP

Change-Id: I49e36918bc231c08e1a473fe99e5086f54416850