Limit transient images tests
Transient images (images that usage field contains the flag bit
VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT) can be used only in
conjunction with following usages (according to specification):
 * VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT
 * VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT
 * VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT
The tests falsely attempts to create images with layout:
 * VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL
 * VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL
Such layout requires images to be created with usages:
 * VK_IMAGE_USAGE_TRANSFER_SRC_BIT
 * VK_IMAGE_USAGE_TRANSFER_DST_BIT
that's prohibited for transient images.

If underlying image is transient, it is invalid for it to be a
TRANSFER_SRC or TRANSFER_DST. Do not allow tests for transient
images to use such layouts.

Affected tests:
 * dEQP-VK.renderpass.*.attachment.*
 * dEQP-VK.renderpass2.*.attachment.*

Components: vulkan

VK-GL-CTS issue: 1545

Change-Id: Ibd1635d15419adcb646fbc1daf030a2736b972b1