Fix image sizes for formats that need a multiple of 2.
All these have in their format definition something like

"Images in this format must be defined with a width that is a multiple
 of two."

in their format definition.

This patch makes sure we actually use even sizes for images with these
formats.

It is not ideal that I have to hardcode these lists but I don't know a
better source to get them from.

Affects:

dEQP-VK.api.invariance.random

Components: Vulkan

VK-GL-CTS issue: 1766

Change-Id: Ie19583db56e7ae483136e1fb4651ae625efb9c02