Add SPIR-V unused variable tests
These tests check everything works when a shader contains references to
variables that have a binding which is not present in the pipeline
layout. For the test to be legal, the variable cannot be used from the
entry point. The tests have two variants: either the variable is simply
unused or the variable is used by a function that is not part of the
entry point call tree.

New tests:
dEQP-VK.spirv_assembly.instruction.*.unused_variables.*

Components: Vulkan
VK-GL-CTS issue: 1038

Change-Id: Ib54c79da4d164e494c0004f0cb1503783e38bc70