Add tests for VK_EXT_pipeline_creation_feedback extension
Tests implement the following for both compute and graphics pipelines.
They create a pipeline cache and checks the pipeline creation feedback
for three different pipelines: first pipeline should not hit the cache,
second pipeline is the same than the first one, third pipeline is a
pipeline derivative from the first one.

Then they verify that the creation feedback for all of the pipelines is
valid, the first pipeline did not hit the cache and that first and
second pipelines are not setting
VK_PIPELINE_CREATION_FEEDBACK_BASE_PIPELINE_ACCELERATION_BIT_EXT as
they are not pipeline derivatives.

Also, if pPipelineStageCreationFeedbacks for each pipeline and stage is
not valid, check that none of the flags is set.

They also print to the test log the content of the
pPipelineCreationFeedback for each pipeline and, if
pPipelineStageCreationFeedbacks for each pipeline and stage is valid,
its content too.

New tests:

dEQP-VK.pipeline.creation_feedback.*

Components: Vulkan
VK-GL-CTS issue: 1671

Change-Id: I5e8b728bd83c256b3b4f084c6d0840a132afbcae